accessibility issue fix
This is a solution for an form label missing for the search.  The drop downs to the right also still need ids, and labels attached to them as well, but will need someone better than i am at php. format should look the same as the end result of the search i just added .